Added new example, a -c option to rbc_make, and made rbc_make call the preprocessing code directly.

+++ rbc_pp.rb
@@ -7,10 +7,6 @@
 #  author: gordon.sommers@gmail.com
 #
 
-
-
-require 'pp'
-
 class Integer
   def even?
     self % 2 == 0
@@ -20,10 +16,9 @@ class Integer
   end
 end
 
-
-def parse_file(text, start, stop, out)
+def parse_file(in_fh, out, start="$START$", stop="$STOP$")
   re = /#{Regexp.quote(start)}(.+?)#{Regexp.quote(stop)}/m;
-  source_blocks = text.split(re)
+  source_blocks = in_fh.read.split(re)
   source_blocks.each_index do |i|
     out.print source_blocks[i] if i.even?
     out.print eval(source_blocks[i]) if i.odd?
@@ -39,33 +34,28 @@ def usage
 END
 end
 
-xloc = ARGV.index("-X")
-floc = nil
-fname = nil
-start = "$START$"
-stop = "$STOP$"
-if (xloc)
-  if (xloc + 2 >= ARGV.length) then usage; exit 1; end
-  start = ARGV[xloc+1]
-  stop = ARGV[xloc+2]
-  if (FileTest.exist?(start) || FileTest.exist?(stop))
-    STDERR.puts "*WARN*: you specified a filename as your START or STOP arg\n*WARN*:  are you sure you passed your arguments in the right order?"
+if __FILE__ == $0
+  xloc = ARGV.index("-X")
+  floc = nil
+  fname = nil
+  start = "$START$"
+  stop = "$STOP$"
+  if (xloc)
+    if (xloc + 2 >= ARGV.length) then usage; exit 1; end
+    start = ARGV[xloc+1]
+    stop = ARGV[xloc+2]
+    if (FileTest.exist?(start) || FileTest.exist?(stop))
+      STDERR.puts "*WARN*: you specified a filename as your START or STOP arg\n*WARN*:  are you sure you passed your arguments in the right order?"
+    end
+    floc = (xloc + 3 >= ARGV.length) ? xloc - 1 : xloc + 3
+    fname = (floc >= 0) ? ARGV[floc] : nil if floc
+  else
+    fname = ARGV[0] if ARGV[0]
   end
-  floc = (xloc + 3 >= ARGV.length) ? xloc - 1 : xloc + 3
-  fname = (floc >= 0) ? ARGV[floc] : nil if floc
-else
-  fname = ARGV[0] if ARGV[0]
+
+  f = (fname) ? File.open(fname) : STDIN
+  parse_file(f, STDOUT, start, stop)
+  f.close if fname #if we used a file, close it
 end
-f = (fname) ? File.open(fname) : STDIN
-
-###
-text = f.read
-parse_file(text, start, stop, STDOUT)
-###
-
-f.close if fname #if we used a file, close it
-
-
-
 ##### ##### ##### ##### #####
